add ngtt

+++ ngtt/uplink/connection.py
@@ -4,9 +4,9 @@ import typing as tp
 import os
 import socket
 import tempfile
-from ssl import SSLContext, PROTOCOL_TLS_CLIENT, SSLError
+from ssl import SSLContext, PROTOCOL_TLS_CLIENT, SSLError, CERT_REQUIRED
 
-from satella.coding import silence_excs, rethrow_as
+from satella.coding import silence_excs, rethrow_as, Closeable, wraps
 from satella.coding.concurrent import IDAllocator
 from satella.files import read_in_file
 
@@ -18,10 +18,23 @@ from .certificates import get_device_info, get_dev_ca_cert, get_root_cert
 PING_INTERVAL_TIME = 30
 
 
-class NGTTSocket:
+def must_be_connected(fun):
+    @wraps(fun)
+    def outer(self, *args, **kwargs):
+        if not self.connected:
+            self.connect()
+        return fun(self, *args, **kwargs)
+    return outer
+
+
+class NGTTSocket(Closeable):
+    __slots__ = ('host', 'cert_file', 'key_file', 'socket',
+                 'buffer', 'w_buffer', 'ping_id', 'last_read', 'connected')
+
     def __init__(self, cert_file: str, key_file: str):
-        self.environment = get_device_info(read_in_file(cert_file))[1]
-        self.host = env_to_hostname(self.environment)
+        super().__init__()
+        environment = get_device_info(read_in_file(cert_file))[1]
+        self.host = env_to_hostname(environment)
         self.cert_file = cert_file
         self.key_file = key_file
         self.socket = None
@@ -30,6 +43,7 @@ class NGTTSocket:
         self.ping_id = None
         self.last_read = None
         self.socket.setblocking(False)
+        self.connected = False
 
         with tempfile.NamedTemporaryFile('wb', delete=False) as chain_file:
             chain_file.write(read_in_file(self.cert_file))
@@ -40,7 +54,8 @@ class NGTTSocket:
 
     @rethrow_as(ssl.SSLError, ConnectionFailed)
     @silence_excs(ssl.SSLWantWriteError)
-    def send_frame(self, tid: int, header: NGTPHeaderType, data: bytes):
+    @must_be_connected
+    def send_frame(self, tid: int, header: NGTPHeaderType, data: bytes = b'') -> None:
         """
         Schedule a frame to be sent
 
@@ -48,13 +63,16 @@ class NGTTSocket:
         :param header: packet type
         :param data: data to send
         """
-        self.w_buffer.extend(STRUCT_LHH.pack(len(data, tid, header.value)))
+        if self.closed:
+            return
+        self.w_buffer.extend(STRUCT_LHH.pack(len(data), tid, header.value))
         self.w_buffer.extend(data)
         data_sent = self.socket.send(self.w_buffer)
         del self.w_buffer[:data_sent]
 
     @rethrow_as(ssl.SSLError, ConnectionFailed)
     @silence_excs(ssl.SSLWantWriteError)
+    @must_be_connected
     def try_send(self):
         """
         Try to send some data
@@ -66,11 +84,13 @@ class NGTTSocket:
             except socket.timeout:
                 return
 
+    @must_be_connected
     def try_ping(self):
         if time.monotonic() - self.last_read > PING_INTERVAL_TIME and self.ping_id is None:
             self.ping_id = self.id_assigner.allocate_int()
             self.send_frame(self.ping_id, NGTPHeaderType.PING, b'')
 
+    @must_be_connected
     def got_ping(self):
         if self.ping_id is not None:
             self.id_assigner.mark_as_free(self.ping_id)
@@ -81,10 +101,12 @@ class NGTTSocket:
 
     @rethrow_as(ssl.SSLError, ConnectionFailed)
     @silence_excs(ssl.SSLWantReadError)
+    @must_be_connected
     def recv_frame(self) -> tp.Optional[tp.Tuple[int, NGTPHeaderType, bytes]]:
         """
         Receive a frame from remote socket
 
+        :raises ConnectionFailed: connection closed
         :return: a tuple of transaction ID, header type, data
         """
         self.try_send()
@@ -94,15 +116,17 @@ class NGTTSocket:
         self.last_read = time.monotonic()
         self.buffer.extend(data)
         if len(self.buffer) > STRUCT_LHH.size:
-            tid, h_type, length = STRUCT_LHH.unpack(self.buffer[:STRUCT_LHH.size])
+            length, tid, h_type = STRUCT_LHH.unpack(self.buffer[:STRUCT_LHH.size])
             if len(self.buffer) < STRUCT_LHH.size + length:
-                return None
+                return
             data = self.buffer[STRUCT_LHH.size:STRUCT_LHH.size+length]
             del self.buffer[:STRUCT_LHH.size+length]
             return tid, NGTPHeaderType(h_type), data
 
-    def __del__(self):
-        os.unlink(self.chain_file_name)
+    def close(self):
+        if super().close():
+            self.disconnect()
+            os.unlink(self.chain_file_name)
 
     def disconnect(self):
         """
@@ -121,8 +145,9 @@ class NGTTSocket:
         if self.socket is not None:
             return
         ssl_context = SSLContext(PROTOCOL_TLS_CLIENT)
-        ssl_context.load_verify_locations(cadata=get_root_cert().encode('utf-8'))
+        ssl_context.load_verify_locations(cadata=get_root_cert().decode('utf-8'))
         ssl_context.load_cert_chain(self.chain_file_name, self.key_file)
+        ssl_context.verify_mode = CERT_REQUIRED
         sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
         ssl_sock = ssl_context.wrap_socket(sock, server_hostname=self.host)
         try:
@@ -132,3 +157,6 @@ class NGTTSocket:
         except (socket.error, SSLError):
             raise ConnectionFailed()
         self.last_read = time.monotonic()
+        self.buffer = bytearray()
+        self.w_buffer = bytearray()
+        self.connected = True
